Add implementation and tests of Archer tool
The tool provides TSAN annotations for OpenMP synchronization. Enable for
application execution by setting OMP_TOOL_LIBRARIES=libarcher.so, this
library should get installed in the same order as libomp, so the linker
should find the file when the application is linked with OpenMP.

Further it is recommended to set
TSAN_OPTIONS="ignore_noninstrumented_modules=1"
this avoids some false positive errors from the OpenMP runtime.